TropCyclone: vectorize and use geosphere metric
Add geodistance function `dist_approx` to util.coordinates and use this instead of
the one in util.interpolation. The new function is able to compute distances in a
spherical metric (with vectorized haversine formula). The @static_method is removed from
`TropCyclone._tc_from_track`. The function `_windfield` is fully vectorized (without numba).