Coerce user-supplied Centroid lat-lon data to np.arrays
This is a small edit to coerce Centroid.lat and Centroid.lon to
np.array objects when they're directly supplied by the user.

Previously a Centroid class with user-supplied lat and lon arrays
(Centroid.set_lat_lon) preserves whatever lat/lon class is passed to it.
This is fine in most cases, but if it's given a pandas.Series,
the TropCyclone.set_from_tracks() method falls  over because it's
unable to use np.nonzero() on a Series.