Avoid some errors by coercing some Impact.select() input vars (#208)
* Avoid some errors by coercing some Impact.select() input vars

Impact.select() takes lists as input for its dates, event_ids and
event_names parameters. If you instead pass a numpy array, the
function fails at line 1031:
   if (dates, event_ids, event_names) != (None, None, None):

with a ValueError:
  ValueError: The truth value of an array with more than one
    element is ambiguous. Use a.any() or a.all()

It's not obvious to the user why this is. So I've added lines at the
start of this method to coerce the input to a list. This will allow
a bit more flexibility with inputs a when it fails it will be more
obvious why.

* Change list(x) to x.tolist() in previous commit

* Impact.select: Use numpy array operations and fix empty selection

Co-authored-by: Thomas Vogt <thomas.vogt@pik-potsdam.de>