Fix Impact's _build_exp_event to work with ids indexing from  > 0
Impact's _build_exp_event is used in a couple of the plotting
methods. It takes an event ID as input and constructs an
exposure object at Impact's centroids, with the impact as
the exposure values, and calls the relevant exposure plotting
method.

This fixes how it deals with the event ID it's provided:
previously it assumed the ID was also the integer position of
that event in the Impact.event_id array. The fix explicitly
finds its position.