Remove unreachable code in Measure class (#281)
I assume this is a leftover from the grand if -> impf renaming. The Measure.calc_impact
method has a return statement part-way through, with the remainder of the code a repeat
of the earlier code, only with an intermediate variable renamed from new_ifs to new_impfs.

I assume it's safe to drop this part!