Extend Exposures.concat to assign new category_ids during concatenation
When combining multiple exposures, you can now supply the `category_ids`
 parameter, a list that sets the category_id of each input Exposures to
the supplied value before concatenating.

This can be used to keep track of source data, or (in my use case) as a
way to quickly read in and combine data on fractional land use types
where each input Exposure is a different land use.