Update coordinates.write_raster to write multiple bands
The points_to_raster method that write_raster calls is designed to write multiple
bands to a file. This makes write_raster able to take advantage of that without
messing with backward compatibility.

The commit also updates and comments a logic check in the method. The check is there
 to avoid an unnecessary call to points_to_raster when the Exposures object already
contains complete gridded data. The check was outdated since the existence of a meta
property in an Exposures no longer implies the existence of meta.height and meta.width
properties (since now all Exposures are created with an Exposures.meta.crs property).

This also updates the documentation.