Make stricter check when using  raster approach to assign centroids
In `Exposures.assign_centroids`, the first approach the method tries
involves rasters, since this is fast. It checks to see if the
Hazard's Centroids have a `meta` property to decide if this is
appropriate.

However a Centroids having a meta isn't equivalent to being a raster
grid. For example, `Centroids.union` sets a meta, as does calling
`Hazard.lat_lon_to_meta`.

This check makes sure that the number of centroids matches the
number of grid cells defined by the meta (and that the meta defines
this).