Add Coverage tests of exposures/base.py
2 Coverage tests added on climada/entity/exposure/test/test.base.py
The line tested are line 171 and 914 of climada/entity/exposure/base.py

The test for Line 171 check if an error is raised when the __init__
function is passed with meta of type list, and also test if the error
raised is the correct one.

The test for Line 914 check if an error is raised when both crs and epsg
are passed to the function to_crs, and also check if the error
raised is the correct one.