Add new coverage tests
Add 2 new test for the file climada/entity/exposures/test/test.base.py
The test check functions at line 171 and 914 of the file:
climada/entity/exposures/base.py.

1) The first test (line 171): Check if a ValueError is raised when meta
of type list in passed to the function __init__. It also check if the
ValueError raised is the excat same as expected.

2) Second test (line 914): Check if a ValueError is raised when both
crs and epsg are passed to the function to_crs. It also check if the
ValueError raised is the excat same as expected