Fix bug in tutorial
Fix bug in the tutorial climada_entity_Impact_FuncSet.

Unfortunately, not all the outputs could be reverted after the changes
were saved. Therefore, a lot of outputs have changed (plots mainly).

The arguments that define an impact function where not
specified with keywords when Impct.Func were called
to initialize an arbitrarly imp_fun, additionaly there where in
a different order in respect to the __init__ function of ImpactFunc.
Therefore, when calling Impact.Func.mdd, Impact.Func.paa and
Impact.Func.intensity, the following wrong outputs were produced.

imp_fun.intensity was instanciated as intensity units
imp_fun.intensity_unit was instanciated as mdd
imp_fun.name  was instanciated as paa
imp_fun.haz_type  was instanciated as id
imp_fun.paa  was instanciated as intesity

Therefore when u_check was called to check that the size of the
intensity is equal to the size of mdd and paa, it was not comparing
the actual real intesities, mdd and paa, but the above values.

By specifyng the arguments as in the next paragraph, u_check,
situtated in: climada/entity/impact_funcs/base.py, compare that
the size of the real intensity is equal to the size of the real mdd
and paa.

Specification of the arguments with keyword:

imp_fun = ImpactFunc(id=id, name=name, intensity_unit=intensity_unit,
 haz_type=haz_type, intensity=intensity, mdd=mdd, paa=paa)