TC synth: update high lat decay, set target pressure with only sea frames, bugfix landfall index
- When creating synthetic tracks, we apply decays to storms at very high or
very low latitudes that are at much higher intensities than anything in the
historic record. This updates the parameters in tc_tracks_synth so that at
tracks always decay when they get beyond the historically observed latitudes
in that basin
- Add the track modified triggered by the env variable
TRACKGEN_TEST_TARGET_PRESSURE_OVER_SEA. When a synthetic track is choosing a
target minimum pressure it now uses only frames from the original track that
are over the ocean. This prevents it from choosing the first frame after
landfall which can be an unintentionally intense value
- Change a few of the timing checks to be based on time rather than number of
frames, since the time step can't be assumed to be six hours
- Update a landfall index calculation which which chose an index off by one
- Add more information while debugging on the tracks that are changed or
removed