"Bugfix: don't use small distances as denominators"
As part of the track extension, the required size of a vector is
estimated using the distance since landfall. When this is tiny a
huge vector is allocated, occasionally causing a memory error.

This puts a minimum value of 1 km on that timestep which still
allows for large vectors to be allocated, just not infinite ones.