Print relevant env info when loading tc_tracks_synth code
Some of the settings in the new TCTracks synthetic generation are read
from environment variables (at least for now while we're validating).

Since these are read when the package is attached, and not when the code
is run, it won't see any variables that are changed after the module
import and this has caused me problems. So for info and transparency the
logger prints information about these settings as it reads them.