Refine Risk Assessment Terms: Vulnerability and Impact Functions (Update Guide_Introduction.ipynb) (#802)
* Update Guide_Introduction.ipynb

The choice of terminology in risk assessment models such as CLIMADA is crucial, as it shapes the framework and perspective of the analysis. The distinction between "vulnerability function" and "impact function" is particularly significant. I made some changes along the lines of making this terminology more visible. I also made changes in the documentation where vulnerability function or damage function was still referenced in order to unify the communication of CLIMADA terminology. 

I considered adding a few lines to explain that the concepts of damage function, vulnerability function, and impact function, though interrelated, each serve a distinct purpose. Damage functions quantify specific damages, focusing on concrete impacts. Vulnerability functions, in contrast, articulate the susceptibility of different exposures to hazards. This susceptibility encompasses more than just the potential for damage; it includes a variety of preconditions that define an exposure's overall vulnerability. Meanwhile, impact functions assess the likelihood of a hazard's effects, both negative and positive, by parameterizing the degree to which an exposure will be impacted by a particular hazard.Something along those lines..... but maybe it's too much... we can discuss it sometime ......

* Update Guide_Introduction.ipynb

* Implementations of comments